requirements.txt
# GUI
PySide6
qtawesome

# Audio & YouTube
yt-dlp
pydub
librosa
soundfile

# Whisper (transkrypcja)
git+https://github.com/openai/whisper.git

# Eksport plików
python-docx
srt